FIX - Readline Problems With Ruby on OSX -
The relevant error message and failure, issued by byebug was:

Sorry, you can't use byebug without Readline. To solve this, you need to rebuild Ruby with Readline support. If using Ubuntu, try `sudo apt-get install

For a longer-term fix we are experimenting with including the rb-readline gem in the development group of our Gemfile.